app/requirements.txt
aiohttp==3.9.3
fastapi==0.103.2
python-dotenv==1.0.1
uvicorn==0.23.2
vertexai==1.43.0
google-cloud-logging==3.6.0
slack-bolt==1.18.0
slack-sdk==3.23.0
slackclient==2.9.4
slackeventsapi==3.0.1
